+++ bitstring/bitstring_test.go
@@ -1,7 +1,6 @@
 package bitstring
 
 import (
-	"fmt"
 	"math/rand"
 	"testing"
 
@@ -345,82 +344,4 @@ func TestBitStringSwapSubstring(t *testing.T) {
 				"want %s, got %s", tt.expZeros, zeros.String())
 		})
 	}
-}
-
-func ExampleNew() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	// upon creation all bits are unset
-	fmt.Println(bitstring)
-	// Output: 00000000
-}
-
-func ExampleNewFromString() {
-	// create a BitString from string
-	bitstring, _ := MakeFromString("101001")
-	fmt.Println(bitstring)
-	// Output: 101001
-}
-
-func ExampleBitString_Len() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	fmt.Println(bitstring.Len())
-	// Output: 8
-}
-
-func ExampleBitString_Bit() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	fmt.Println(bitstring.Bit(7))
-	// Output: false
-}
-
-func ExampleBitString_SetBit() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	bitstring.SetBit(2, true)
-	fmt.Println(bitstring)
-	// Output: 00000100
-}
-
-func ExampleBitString_FlipBit() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	bitstring.FlipBit(2)
-	fmt.Println(bitstring)
-	// Output: 00000100
-}
-
-func ExampleBitString_ZeroesCount() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	// upon creation all bits are unset
-	fmt.Println(bitstring.ZeroesCount())
-	// Output: 8
-}
-
-func ExampleBitString_OnesCount() {
-	// create a 8 bits BitString
-	bitstring, _ := New(8)
-	// upon creation all bits are unset
-	fmt.Println(bitstring.OnesCount())
-	// Output: 0
-}
-
-func ExampleBitString_BigInt() {
-	// create a 8 bits BitString
-	bitstring, _ := MakeFromString("100")
-	bi := bitstring.BigInt()
-	fmt.Println(bi.Int64())
-	// Output: 4
-}
-
-func ExampleBitString_SwapSubstring() {
-	bs1, _ := MakeFromString("111")
-	bs2, _ := MakeFromString("000")
-	// starting from bit 2 of bs1, swap 1 bit with bs2
-	bs1.SwapRange(bs2, 2, 1)
-	fmt.Println(bs1)
-	// Output: 011
 }
